services/federated-aggregator/requirements.txt
numpy>=1.24.0
structlog>=23.1.0
google-cloud-pubsub>=2.18.0
google-cloud-firestore>=2.13.0
pydantic>=2.5.0
